requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.30.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.0
google_search_results==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.6
langchain-anthropic==0.1.16
langchain-community==0.2.6
langchain-core==0.2.10
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.82
load-dotenv==0.1.0
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.10.1
numpy==1.26.4
orjson==3.10.5
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.4.2
tokenizers==0.19.1
tqdm==4.66.4
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
